added projectiles

+++ src/Level.java
@@ -122,8 +122,8 @@ public class Level {
      * @param y Vertical map coordinate.
      * @return Tile at parsed position.
      */
-    public Tile getTile(int x, int y) {
-        return tiles[x][y];
+    public Tile getTile(double x, double y) {
+        return tiles[(int)x][(int)y];
     }
 
     /**
@@ -151,10 +151,10 @@ public class Level {
      * @param y int y value that shares the tile
      * @return Arraylist elements contained in the tile that has coordinates x,y
      */
-    public ArrayList<Element> getElements(int x, int y) {
+    public ArrayList<Element> getElements(double x, double y) {
         ArrayList<Element> stack = new ArrayList<>(); //returning elements at x,y
         for (Element element : elements) {
-            if (element.getX() == x && element.getY() == y) {
+            if (element.getX() == (int)x && element.getY() == (int)y) {
                 stack.add(element);
             }
         }
